requirements.txt
# Pydantic for type-safe configuration management
pydantic
# QT bindings for GUI
PySide6
PyYAML